+++ src/code/SecretStore.cs
@@ -56,23 +56,55 @@ namespace Microsoft.PowerShell.SecretStore
     /// <summary>
     /// Updates the local store password to the new password provided.
     /// </summary>
-    [Cmdlet(VerbsCommon.Set, "SecretStorePassword")]
+    [Cmdlet(VerbsCommon.Set, "SecretStorePassword", DefaultParameterSetName = NoParameterSet)]
     public sealed class SetSecretStorePasswordCommand : PSCmdlet
     {
+        #region Members
+
+        private const string NoParameterSet = "NoParameterSet";
+        private const string ParameterSet = "ParameterSet";
+
+        #endregion
+
+        #region Parameters
+
+        [Parameter(Position=0, Mandatory=true, ParameterSetName=ParameterSet, ValueFromPipeline=true)]
+        [ValidateNotNull]
+        public SecureString NewPassword { get; set; }
+
+        [Parameter(Position=1, ParameterSetName=ParameterSet)]
+        public SecureString Password { get; set; }
+
+        #endregion
+
         #region Overrides
 
         protected override void EndProcessing()
         {
             SecureString newPassword;
             SecureString oldPassword;
-            oldPassword = Utils.PromptForPassword(
-                cmdlet: this,
-                verifyPassword: false,
-                message: "Old password");
-            newPassword = Utils.PromptForPassword(
-                cmdlet: this,
-                verifyPassword: true,
-                message: "New password");
+
+            switch (ParameterSetName)
+            {
+                case NoParameterSet:
+                    oldPassword = Utils.PromptForPassword(
+                        cmdlet: this,
+                        verifyPassword: false,
+                        message: "Old password");
+                    newPassword = Utils.PromptForPassword(
+                        cmdlet: this,
+                        verifyPassword: true,
+                        message: "New password");
+                    break;
+
+                case ParameterSet:
+                    oldPassword = Password;
+                    newPassword = NewPassword;
+                    break;
+
+                default:
+                    throw new InvalidOperationException("Unknown parameter set");
+            }
 
             LocalSecretStore.GetInstance(password: oldPassword).UpdatePassword(
                 newPassword,
@@ -239,6 +271,12 @@ namespace Microsoft.PowerShell.SecretStore
         public Interaction Interaction { get; set; }
 
         [Parameter]
+        public SecureString Password { get; set; }
+
+        [Parameter]
+        public SwitchParameter PassThru { get; set; }
+
+        [Parameter]
         public SwitchParameter Force { get; set; }
 
         #endregion
@@ -275,11 +313,12 @@ namespace Microsoft.PowerShell.SecretStore
             }
 
             var defaultConfigData = SecureStoreConfig.GetDefault();
+            var interaction = MyInvocation.BoundParameters.ContainsKey(nameof(Interaction)) ? Interaction : defaultConfigData.Interaction;
             var newConfigData = new SecureStoreConfig(
                 scope: MyInvocation.BoundParameters.ContainsKey(nameof(Scope)) ? Scope : defaultConfigData.Scope,
                 authentication: MyInvocation.BoundParameters.ContainsKey(nameof(Authentication)) ? Authentication : defaultConfigData.Authentication,
                 passwordTimeout: MyInvocation.BoundParameters.ContainsKey(nameof(PasswordTimeout)) ? PasswordTimeout : defaultConfigData.PasswordTimeout,
-                interaction: MyInvocation.BoundParameters.ContainsKey(nameof(Interaction)) ? Interaction : defaultConfigData.Interaction);
+                interaction: interaction);
 
             if (!SecureStoreFile.RemoveStoreFile(out string errorMsg))
             {
@@ -305,7 +344,24 @@ namespace Microsoft.PowerShell.SecretStore
 
             LocalSecretStore.Reset();
 
-            WriteObject(newConfigData);
+            if (Password != null)
+            {
+                LocalSecretStore.GetInstance(
+                    password: Password).UnlockLocalStore(
+                        password: Password,
+                        passwordTimeout: MyInvocation.BoundParameters.ContainsKey(nameof(PasswordTimeout)) ? 
+                            (int?)PasswordTimeout : null);
+            }
+            else if (interaction == Microsoft.PowerShell.SecretStore.Interaction.Prompt)
+            {
+                // Invoke the password prompt.
+                LocalSecretStore.GetInstance(cmdlet: this);
+            }
+
+            if (PassThru.IsPresent)
+            {
+                WriteObject(newConfigData);
+            }
         }
 
         #endregion