+++ CHANGELOG.md
@@ -6,11 +6,17 @@
 
 ### Changes
 
-- `Set-SecretStoreConfiguration` now has a `-PassThru` parameter to write the store configuration object to the pipeline, and no longer writes the configuration object by default (Issue #25).
+- `Set-SecretStoreConfiguration` now has a `-PassThru` parameter to write the store configuration object to the pipeline, and no longer writes the configuration object by default (Issue #29).
 
 - A `-PasswordTimeout` value of zero now allows the provided password to be used only once (Issue #30).
 
-- When setting a password, an empty password is no longer accepted (Issue #31)
+- When setting a password, an empty password is no longer accepted (Issue #31).
+
+- `Set-SecretStorePassword` now has a parameter set that takes old/new password arguments, to allow setting password by automation (Issue #26).
+
+- `Reset-SecretStore` now has a new `-Password` and `-PassThru` parameter.
+
+- `Reset-SecretStore` will now prompt immediately for a new password, if password authentication is selected and prompt interaction is allowed.
 
 ### New Features